Assistant checkpoint: Update UI Simplification project page with new content
Assistant generated file changes:
- client/src/pages/work/utopia/UiSimplification.tsx: Update UI Simplification project content with new description

---

User prompt:

update the UISimplification.tsx file to use my project description. make new paragraphs and headings as needed. Also leave sample images as needed, i will replace them with the real images later. Also please proofread my work and make small tweaks if necessary to optimize my portfolio to be professional, yet fun and friendly and creative -

There came a point in the development of Utopia when we realized the style of the editor UI needed a fresh look. The panels were too close together, stretched to the edges of the screen, and the toolbar had way too many buttons and labels. It felt like a tool for experts, and we wanted to make the explorer feel welcome here too. We decided to redesign the UI to be more fluid, and allow users to customize their workspace.

[image/gif]

Utopia is for everyone, but not everyone works the same way. With this simplified UI, you can reorder, resize, and close each panel however it is most helpful for you. Don’t want to see the code editor? No problem, close it. Or maybe you want to see the code editor, but you aren’t really using it. Also great! You can have a minified view of it, in any corner you like. Or maybe you’re doing some heavy software developing and you need it full size. You can do that too.

Inspiration

We took inspiration from contemporary tools like Spline and tldraw for their style of floating panels. We also made all of our UI panels appear to “float” above the canvas. This was the biggest visual change, and gives the UI the fresh modern feel we were going for.

Similar to Framer, we decided to use the toolbar as a place to surface certain important messages like “View Only” when you don’t have access to edit a project. We also wanted to use the toolbar as a way of organizing the different “modes” you could play with. Today we have edit, text, insert, play, and comment mode. Each of which trigger their own respective cursors.

[images]

Challenges

Implementing this concept was not as easy as one would hope! It took multiple engineers several attempts at finding an appropriate development solution that would let the users drag to reorder and resize the panels, in a way that still felt intuitive.

Thanks to Balazs and Eni, some of the talented engineers on the team, we eventually landed on the solution we use today. Essentially there is an invisible grid above the canvas, that acts as the container for the floating panels. We use the react drag and drop library to enable the reordering.

Resulting Experience

Before [image]
After [image]